refactor(parallel): isolate pool behind feature
