+++ Cargo.toml
@@ -35,6 +35,7 @@ htslib = ["dep:rust-htslib"]
 url = ["dep:reqwest", "dep:niffler"]
 ssh = ["dep:which"]
 gcs = ["dep:which"]
+pool = []
 
 [dev-dependencies]
 clap = { version = "4.5.40", features = ["derive"] }
