+++ src/parallel/mod.rs
@@ -2,6 +2,7 @@ mod error;
 pub(crate) mod multi;
 mod ordered;
 pub(crate) mod paired;
+#[cfg(feature = "pool")]
 pub mod pool;
 mod processor;
 pub(crate) mod reader;
@@ -12,4 +13,7 @@ pub use ordered::Ordered;
 pub use processor::{MultiParallelProcessor, PairedParallelProcessor, ParallelProcessor};
 pub use reader::ParallelReader;
 
+#[cfg(feature = "pool")]
 pub use pool::ThreadPool;
+#[cfg(feature = "pool")]
+pub use reader::PoolParallelReader;
